Add logic to prevent a race-condition during simultaneous OOM restarts

+++ sarracenia/flowcb/housekeeping/resources.py
@@ -34,7 +34,7 @@ Returns:
 
 import logging
 
-import os, socket
+import os, socket, time, random
 from sarracenia.flowcb import FlowCB
 from sarracenia import naturalSize, naturalTime, user_cache_dir, nowstr
 from sarracenia.featuredetection import features
@@ -58,6 +58,7 @@ class Resources(FlowCB):
         ''' Per-process maximum memory footprint that is considered too large, forcing a process restart.'''
         self.transferCount = 0
         self.msgCount = 0
+        self.randomSleep = 1 + round(random.random(), 2)
 
     def on_housekeeping(self):
         if features['process']['present']:
@@ -117,8 +118,17 @@ class Resources(FlowCB):
         # Before triggering a restart, add a state file to prevent other processes to stop/start it at the same time.
         self.state_file = self.o.cfg_run_dir + os.sep + 'resources_restart'
 
-        with open(self.state_file, "w") as f:
-            f.write(nowstr())
+        file_not_ready = True
+
+        # If another process is performing a OOM restart, wait until it is complete before creating a new state file to avoid a race-condition.
+        while file_not_ready:
+            if not os.path.exists(self.state_file):
+                file_not_ready = False
+                with open(self.state_file, "w") as f:
+                    f.write(nowstr())
+            else:
+                logger.info(f"State file already exists : {self.state_file}. Waiting {self.randomSleep} seconds.")
+                time.sleep(self.randomSleep)
 
 
         if sys.platform.startswith(('linux', 'cygwin', 'darwin', 'aix')):
